fix: combines functions into a component filter.js
