+++ src/components/header/filter.js
@@ -1,7 +1,24 @@
-import {createElement} from "../../../helpers/utils.js";
-import {createFilterMarkup} from './filter-markup';
+import {createElement} from "../../helpers/utils.js";
 
 const FILTER_NAMES = [`everything`, `future`, `past`];
+
+const createFilterMarkup = (name, isChecked) => {
+  const checked = `${isChecked ? `checked` : ``}`;
+
+  return (
+    `<div class="trip-filters__filter">
+      <input
+        id="filter-${name}"
+        class="trip-filters__filter-input  visually-hidden"
+        type="radio"
+        name="trip-filter"
+        value="${name}"
+        ${checked}
+      />
+      <label class="trip-filters__filter-label" for="filter-${name}">${name}</label>
+    </div>`
+  );
+};
 
 const createFilterTemplate = () => {
   const filtersMarkup = FILTER_NAMES.map((it, i) => createFilterMarkup(it, i === 0)).join(`\n`);
